Improve spreadsheet export
Only users marked with an admin flag have access to export spreadsheet functionality. The administrator chooses which email(s) he or she wants to be exported. In other news, cleaned up Firebase database rules, and updated some basic CSS. The next step is to check that 3rd party cookies are enabled on the front end, add a button to allow the user to go to the student page at any time, and allow the user to see which email he or she is signed into at the top right of the reading interface.